pi/requirements.txt
Adafruit-Blinka==4.5.0
adafruit-circuitpython-busdevice==4.3.1
adafruit-circuitpython-motor==3.1.1
adafruit-circuitpython-motorkit==1.5.1
adafruit-circuitpython-pca9685==3.3.1
adafruit-circuitpython-register==1.8.1
Adafruit-PlatformDetect==2.6.1
Adafruit-PureIO==1.1.4
astroid==2.3.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pyftdi==0.50.1
pylint==2.4.4
pyserial==3.4
pyusb==1.0.2
rpi-ws281x==4.2.3
RPi.GPIO==0.7.0
six==1.14.0
sysv-ipc==1.0.1
typed-ast==1.4.1
wrapt==1.11.2
